Added user login/logout and register functionality(User Authentication)

+++ ToDo/routers/todos.py
@@ -39,24 +39,41 @@ def get_db():
 
 @router.get('/', response_class=HTMLResponse)
 async def get_all_by_user(request: Request, db: Session = Depends(get_db)):
-    todos = db.query(models.ToDo).all()
-    return templates.TemplateResponse("home.html", {"request":request, "todos": todos})
+    user = await get_current_user(request)
+
+    if user is None:
+        return RedirectResponse(url="/auth", status_code=status.HTTP_302_FOUND)
+
+    todos = db.query(models.ToDo).filter(models.ToDo.owner_id == user.get("id")).all()
+
+    return templates.TemplateResponse("home.html", {"request":request, "todos": todos, "user": user})
 
 
 @router.get('/add-todo', response_class=HTMLResponse)
 async def add_new_todo(request: Request):
-    return templates.TemplateResponse("add-todo.html", {"request":request})
+
+    user = await get_current_user(request)
+
+    if not user:
+        return RedirectResponse(url="/auth", status_code=status.HTTP_302_FOUND)
+
+    return templates.TemplateResponse("add-todo.html", {"request":request, "user": user})
 
 
 @router.post('/add-todo', response_class=HTMLResponse)
 async def create_todo(request: Request, title: str = Form(...), description: str = Form(...),
                        priority: int = Form(...), db: Session = Depends(get_db)):
+    user = await get_current_user(request)
+
+    if not user:
+        return RedirectResponse(url="/auth", status_code=status.HTTP_302_FOUND)
+
     todo_model = models.ToDo()
     todo_model.title = title
     todo_model.description = description
     todo_model.priority = priority
     todo_model.complete = False
-    todo_model.owner_id = 1
+    todo_model.owner_id = user.get("id")
     db.add(todo_model)
     db.commit()
     return RedirectResponse(url="/todos", status_code=status.HTTP_302_FOUND)
@@ -64,13 +81,23 @@ async def create_todo(request: Request, title: str = Form(...), description: str
 
 @router.get('/edit-todo/{todo_id}', response_class=HTMLResponse)
 async def edit_todo(request: Request, todo_id: int, db: Session = Depends(get_db)):
+    user = await get_current_user(request)
+
+    if not user:
+        return RedirectResponse(url="/auth", status_code=status.HTTP_302_FOUND)
+
     todo_to_update = db.query(models.ToDo).filter(models.ToDo.id == todo_id).first()
-    return templates.TemplateResponse("edit-todo.html", {"request":request, "todo": todo_to_update})
+    return templates.TemplateResponse("edit-todo.html", {"request":request, "todo": todo_to_update, "user": user})
 
 
 @router.post('/edit-todo/{todo_id}', response_class=HTMLResponse)
 async def edit_todo_commit(request: Request, todo_id: int, title: str = Form(...), description: str = Form(...),
                        priority: int = Form(...), db: Session = Depends(get_db)):
+    user = await get_current_user(request)
+
+    if not user:
+        return RedirectResponse(url="/auth", status_code=status.HTTP_302_FOUND)
+
     todo_to_update = db.query(models.ToDo).filter(models.ToDo.id == todo_id).first()
     todo_to_update.title = title 
     todo_to_update.description = description
@@ -84,9 +111,15 @@ async def edit_todo_commit(request: Request, todo_id: int, title: str = Form(...
 @router.get('/delete/{todo_id}')
 async def delete_todo(request: Request, todo_id: int, 
                       db: Session = Depends(get_db)):
-     todo = db.query(models.ToDo).filter(models.ToDo.id == todo_id).filter(models.ToDo.owner_id==1).first()
+     
+     user = await get_current_user(request)
+     if not user:
+        return RedirectResponse(url="/auth", status_code=status.HTTP_302_FOUND)
+     
+     todo = db.query(models.ToDo).filter(models.ToDo.id == todo_id).filter(models.ToDo.owner_id==user.get("id")).first()
      if not todo:
-            return RedirectResponse(url="/todos", status_code=status.HTTP_302_FOUND)
+        return RedirectResponse(url="/todos", status_code=status.HTTP_302_FOUND)
+
      db.query(models.ToDo).filter(models.ToDo.id == todo_id).delete()
      db.commit()
 
@@ -97,6 +130,12 @@ async def delete_todo(request: Request, todo_id: int,
 @router.get('/complete/{todo_id}', response_class=HTMLResponse)
 async def complete_todo(request: Request, todo_id: int, 
                       db: Session = Depends(get_db)):
+
+    user = await get_current_user(request)
+
+    if not user:
+        return RedirectResponse(url="/auth", status_code=status.HTTP_302_FOUND)
+
     todo = db.query(models.ToDo).filter(models.ToDo.id == todo_id).first()
     todo.complete = not todo.complete
     db.add(todo)
